-fix: closer countries list
