+++ backend/src/app/mappers/countries.js
@@ -1,11 +1,5 @@
 const { pagination, idParam } = require('./common');
 
-exports.getCloserCountriesMapper = req => ({
-  ...pagination(req),
-  latitude: req.query.latitude,
-  longitude: req.query.longitude
-});
-
 exports.getCountriesMapper = req => ({
   ...pagination(req),
   name: req.query.name,
